+++ packages/native_pdf_view/lib/src/native_pdf_view.dart
@@ -1,25 +1,23 @@
-import 'dart:ui';
-
-import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
-import 'package:extended_image/extended_image.dart';
+import 'package:native_pdf_view/src/pdf_page_image_provider.dart';
+import 'package:photo_view/photo_view.dart';
+import 'package:photo_view/photo_view_gallery.dart';
 import 'package:native_pdf_renderer/native_pdf_renderer.dart';
 import 'package:synchronized/synchronized.dart';
 
 export 'package:native_pdf_renderer/native_pdf_renderer.dart';
-export 'package:extended_image/extended_image.dart';
 
 part 'native_pdf_controller.dart';
 
-typedef PDFViewPageBuilder = Widget Function(
+typedef PDFViewPageBuilder = PhotoViewGalleryPageOptions Function(
   /// Page image model
-  PdfPageImage pageImage,
+  Future<PdfPageImage> pageImage,
 
-  /// true if page now showed
-  bool isCurrentIndex,
+  /// page index
+  int index,
 
-  /// onDoubleTap Animation
-  AnimationController animationController,
+  /// pdf document
+  PdfDocument document,
 );
 
 typedef PDFViewPageRenderer = Future<PdfPageImage> Function(PdfPage page);
@@ -29,7 +27,7 @@ final Lock _lock = Lock();
 /// Widget for viewing PDF documents
 class PdfView extends StatefulWidget {
   const PdfView({
-    @required this.controller,
+    required this.controller,
     this.onPageChanged,
     this.onDocumentLoaded,
     this.onDocumentError,
@@ -41,13 +39,10 @@ class PdfView extends StatefulWidget {
     this.scrollDirection = Axis.horizontal,
     this.pageSnapping = true,
     this.physics,
+    this.backgroundDecoration = const BoxDecoration(),
     this.loaderSwitchDuration = const Duration(seconds: 1),
-    Key key,
-  })  : assert(pageSnapping != null),
-        assert(controller != null),
-        assert(scrollDirection != null),
-        assert(renderer != null),
-        super(key: key);
+    Key? key,
+  }) : super(key: key);
 
   ///
   final Duration loaderSwitchDuration;
@@ -56,25 +51,25 @@ class PdfView extends StatefulWidget {
   final PdfController controller;
 
   /// Called whenever the page in the center of the viewport changes
-  final void Function(int page) onPageChanged;
+  final void Function(int page)? onPageChanged;
 
   /// Called when a document is loaded
-  final void Function(PdfDocument document) onDocumentLoaded;
+  final void Function(PdfDocument document)? onDocumentLoaded;
 
   /// Called when a document loading error
-  final void Function(Object error) onDocumentError;
+  final void Function(Object error)? onDocumentError;
 
   /// Widget showing when pdf document loading
-  final Widget documentLoader;
+  final Widget? documentLoader;
 
   /// Widget showing when pdf page loading
-  final Widget pageLoader;
+  final Widget? pageLoader;
 
   /// Page builder
   final PDFViewPageBuilder pageBuilder;
 
   /// Show document loading error message inside [PdfView]
-  final Widget Function(Exception error) errorBuilder;
+  final Widget Function(Exception error)? errorBuilder;
 
   /// Custom PdfRenderer options
   final PDFViewPageRenderer renderer;
@@ -85,8 +80,11 @@ class PdfView extends StatefulWidget {
   /// Set to false to disable page snapping, useful for custom scroll behavior.
   final bool pageSnapping;
 
+  /// Pdf widget page background decoration
+  final BoxDecoration? backgroundDecoration;
+
   /// Determines the physics of a [PdfView] widget.
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Default PdfRenderer options
   static Future<PdfPageImage> _render(PdfPage page) => page.render(
@@ -94,77 +92,25 @@ class PdfView extends StatefulWidget {
         height: page.height * 2,
         format: PdfPageFormat.JPEG,
         backgroundColor: '#ffffff',
-      );
-
-  static const List<double> _doubleTapScales = <double>[1.0, 2.0, 3.0];
+      ) as Future<PdfPageImage>;
 
   /// Default page builder
-  static Widget _pageBuilder(
-    PdfPageImage pageImage,
-    bool isCurrentIndex,
-    AnimationController animationController,
-  ) {
-    Animation<double> _doubleTapAnimation;
-    void Function() _animationListener;
-
-    Widget image = ExtendedImage.memory(
-      pageImage.bytes,
-      key: Key(pageImage.hashCode.toString()),
-      fit: BoxFit.contain,
-      mode: ExtendedImageMode.gesture,
-      initGestureConfigHandler: (_) => GestureConfig(
-        minScale: 1,
-        maxScale: 3.0,
-        animationMinScale: .75,
-        animationMaxScale: 3.0,
-        speed: 1,
-        inertialSpeed: 100,
-        inPageView: true,
-        initialScale: 1.0,
-        cacheGesture: false,
-      ),
-      onDoubleTap: (ExtendedImageGestureState state) {
-        final pointerDownPosition = state.pointerDownPosition;
-        final begin = state.gestureDetails.totalScale;
-        double end;
-
-        _doubleTapAnimation?.removeListener(_animationListener);
-
-        animationController
-          ..stop()
-          ..reset();
-
-        if (begin == _doubleTapScales[0]) {
-          end = _doubleTapScales[1];
-        } else {
-          if (begin == _doubleTapScales[1]) {
-            end = _doubleTapScales[2];
-          } else {
-            end = _doubleTapScales[0];
-          }
-        }
-
-        _animationListener = () {
-          //print(_animation.value);
-          state.handleDoubleTap(
-              scale: _doubleTapAnimation.value,
-              doubleTapPosition: pointerDownPosition);
-        };
-        _doubleTapAnimation = animationController
-            .drive(Tween<double>(begin: begin, end: end))
-              ..addListener(_animationListener);
-
-        animationController.forward();
-      },
-    );
-    if (isCurrentIndex) {
-      image = Hero(
-        tag: 'pdf_view' + pageImage.pageNumber.toString(),
-        child: image,
+  static PhotoViewGalleryPageOptions _pageBuilder(
+    Future<PdfPageImage> pageImage,
+    int index,
+    PdfDocument document,
+  ) =>
+      PhotoViewGalleryPageOptions(
+        imageProvider: PdfPageImageProvider(
+          pageImage,
+          index,
+          document.id,
+        ),
+        minScale: PhotoViewComputedScale.contained * 1,
+        maxScale: PhotoViewComputedScale.contained * 3.0,
+        initialScale: PhotoViewComputedScale.contained * 1.0,
+        heroAttributes: PhotoViewHeroAttributes(tag: '${document.id}-$index'),
       );
-    }
-    return image;
-  }
 
   @override
   _PdfViewState createState() => _PdfViewState();
@@ -172,37 +118,31 @@ class PdfView extends StatefulWidget {
 
 class _PdfViewState extends State<PdfView> with SingleTickerProviderStateMixin {
   final Map<int, PdfPageImage> _pages = {};
-  _PdfViewLoadingState _loadingState;
-  Exception _loadingError;
-  int _currentIndex;
-  AnimationController _animationController;
+  late _PdfViewLoadingState _loadingState;
+  Exception? _loadingError;
+  late int _currentIndex;
 
   @override
   void initState() {
-    _animationController = AnimationController(
-      duration: Duration(milliseconds: 200),
-      vsync: this,
-    );
     _loadingState = _PdfViewLoadingState.loading;
     widget.controller._attach(this);
-    _currentIndex = widget.controller._pageController.initialPage ?? 0;
+    _currentIndex = widget.controller._pageController!.initialPage;
     super.initState();
   }
 
   @override
   void dispose() {
     widget.controller._detach();
-    _animationController.dispose();
     super.dispose();
   }
 
   Future<PdfPageImage> _getPageImage(int pageIndex) =>
       _lock.synchronized<PdfPageImage>(() async {
         if (_pages[pageIndex] != null) {
-          return _pages[pageIndex];
+          return _pages[pageIndex]!;
         }
 
-        final page = await widget.controller._document.getPage(pageIndex + 1);
+        final page = await widget.controller._document!.getPage(pageIndex + 1);
 
         try {
           _pages[pageIndex] = await widget.renderer(page);
@@ -210,67 +150,35 @@ class _PdfViewState extends State<PdfView> with SingleTickerProviderStateMixin {
           await page.close();
         }
 
-        return _pages[pageIndex];
+        return _pages[pageIndex]!;
       });
 
   void _changeLoadingState(_PdfViewLoadingState state) {
     if (state == _PdfViewLoadingState.loading) {
       _pages.clear();
     } else if (state == _PdfViewLoadingState.success) {
-      widget.onDocumentLoaded?.call(widget.controller._document);
+      widget.onDocumentLoaded?.call(widget.controller._document!);
     } else if (state == _PdfViewLoadingState.error) {
-      widget.onDocumentError?.call(_loadingError);
+      widget.onDocumentError?.call(_loadingError!);
     }
     setState(() {
       _loadingState = state;
     });
   }
 
-  Widget _buildLoaded() => Listener(
-        onPointerSignal: (PointerSignalEvent event) {
-          if (event is PointerScrollEvent && !widget.pageSnapping) {
-            widget.controller._pageController.animateTo(
-              widget.controller._pageController.offset +
-                  (event.scrollDelta.dy * window.devicePixelRatio),
-              duration: Duration(milliseconds: 75),
-              curve: Curves.ease,
-            );
-          }
+  Widget _buildLoaded() => PhotoViewGallery.builder(
+        builder: (BuildContext context, int index) => widget.pageBuilder(
+            _getPageImage(index), index, widget.controller._document!),
+        itemCount: widget.controller._document!.pagesCount,
+        loadingBuilder: (_, __) => widget.pageLoader ?? SizedBox(),
+        backgroundDecoration: widget.backgroundDecoration,
+        pageController: widget.controller._pageController,
+        onPageChanged: (int index) {
+          _currentIndex = index;
+          widget.onPageChanged?.call(index + 1);
         },
-        child: ExtendedImageGesturePageView.builder(
-          itemBuilder: (BuildContext context, int index) =>
-              FutureBuilder<PdfPageImage>(
-            future: _getPageImage(index),
-            builder: (_, snapshot) {
-              if (snapshot.hasData) {
-                return KeyedSubtree(
-                  key: Key('$runtimeType.page.'
-                      '${widget.controller._document.hashCode}.'
-                      '${_pages[index].pageNumber}'),
-                  child: widget.pageBuilder(
-                    _pages[index],
-                    index == _currentIndex,
-                    _animationController,
-                  ),
-                );
-              }
-
-              return KeyedSubtree(
-                key: Key('$runtimeType.page.loading'),
-                child: widget.pageLoader ?? SizedBox(),
-              );
-            },
-          ),
-          itemCount: widget.controller._document.pagesCount,
-          onPageChanged: (int index) {
-            _currentIndex = index;
-            widget.onPageChanged?.call(index + 1);
-          },
-          controller: widget.controller?._pageController,
-          scrollDirection: widget.scrollDirection,
-          pageSnapping: widget.pageSnapping,
-          physics: widget.physics,
-        ),
+        scrollDirection: widget.scrollDirection,
+        scrollPhysics: widget.physics,
       );
 
   @override
@@ -287,11 +195,8 @@ class _PdfViewState extends State<PdfView> with SingleTickerProviderStateMixin {
       case _PdfViewLoadingState.error:
         content = KeyedSubtree(
           key: Key('$runtimeType.root.error'),
-          child: Padding(
-            padding: EdgeInsets.all(32),
-            child: widget.errorBuilder?.call(_loadingError) ??
-                Center(child: Text(_loadingError.toString())),
-          ),
+          child: widget.errorBuilder?.call(_loadingError!) ??
+              Center(child: Text(_loadingError.toString())),
         );
         break;
       case _PdfViewLoadingState.success: